Feature/chat 보완

+++ src/main/java/com/soongsil/soongpal/chat/service/ChatRoomService.java
@@ -1,5 +1,7 @@
 package com.soongsil.soongpal.chat.service;
 
+import com.soongsil.soongpal.board.domain.Board;
+import com.soongsil.soongpal.board.repository.BoardRepository;
 import com.soongsil.soongpal.chat.domain.ChatRole;
 import com.soongsil.soongpal.chat.domain.ChatRoom;
 import com.soongsil.soongpal.chat.domain.ChatRoomUser;
@@ -20,6 +22,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
+import static com.soongsil.soongpal.chat.domain.ChatRoomType.GROUP;
+import static com.soongsil.soongpal.chat.domain.ChatRoomType.PRIVATE;
+
 
 @Service
 @Transactional
@@ -27,12 +32,44 @@ import java.util.List;
 public class ChatRoomService {
 
     private final UserRepository userRepository;
+    private final BoardRepository boardRepository;
+
     private final ChatRoomRepository chatRoomRepository;
     private final ChatMessageRepository chatMessageRepository;
     private final ChatRoomUserRepository chatRoomUserRepository;
 
-    public ChatRoomResDto createChatRoom(ChatRoomCreateReqDto dto, Long userId) {
-        ChatRoom savedRoom = chatRoomRepository.save(ChatRoomCreateReqDto.toEntity(dto));
+    public ChatRoomResDto createPrivateChatRoom(ChatRoomCreateReqDto dto, Long userId) {
+        ChatRoom savedRoom = chatRoomRepository.save(ChatRoomCreateReqDto.toEntity(dto.getName(), PRIVATE, dto.getBoardId()));
+        User findUser = userRepository.findById(userId)
+                .orElseThrow(() -> new ChatException(ChatErrorCode.USER_NOT_FOUND));
+        Board findBoard = boardRepository.findById(dto.getBoardId())
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시글입니다."));
+        userRepository.findById(findBoard.getUser().getId())
+                .orElseThrow(() -> new ChatException(ChatErrorCode.USER_NOT_FOUND));
+
+        ChatRoomUser roomUser = ChatRoomUser.builder()
+                .chatRoom(savedRoom)
+                .user(findUser)
+                .role(ChatRole.OWNER)
+                .build();
+
+        ChatRoomUser roomOwner = ChatRoomUser.builder()
+                .chatRoom(savedRoom)
+                .user(findBoard.getUser())
+                .role(ChatRole.OWNER)
+                .build();
+
+        chatRoomUserRepository.save(roomOwner);
+        chatRoomUserRepository.save(roomUser);
+
+        savedRoom.addUser(roomOwner);
+        savedRoom.addUser(roomUser);
+
+        return convertToChatRoomResDto(savedRoom);
+    }
+
+    public ChatRoomResDto createGroupChatRoom(Long userId, String chatRoomName, Long boardId) {
+        ChatRoom savedRoom = chatRoomRepository.save(ChatRoomCreateReqDto.toEntity(chatRoomName, GROUP, boardId));
         User findUser = userRepository.findById(userId)
                 .orElseThrow(() -> new ChatException(ChatErrorCode.USER_NOT_FOUND));
 
@@ -60,14 +97,21 @@ public class ChatRoomService {
                 .toList();
     }
 
-    public void joinChatRoom(Long roomId, Long userId) {
-        ChatRoom findChatRoom = chatRoomRepository.findById(roomId)
+    public void joinChatRoom(Long boardId, Long userId) {
+        boardRepository.findById(boardId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시글입니다."));
+
+        ChatRoom findChatRoom = chatRoomRepository.findByBoardId(boardId)
                 .orElseThrow(() -> new ChatException(ChatErrorCode.CHAT_ROOM_NOT_FOUND));
+        
+        if (findChatRoom.getType() == PRIVATE) {
+            throw new ChatException(ChatErrorCode.CHAT_ROOM_ACCESS_DENIED);
+        }
 
         User findUser = userRepository.findById(userId)
                 .orElseThrow(() -> new ChatException(ChatErrorCode.USER_NOT_FOUND));
 
-        boolean alreadyJoined = chatRoomUserRepository.findByChatRoomIdAndUserId(roomId, userId).isPresent();
+        boolean alreadyJoined = chatRoomUserRepository.findByChatRoomIdAndUserId(findChatRoom.getId(), userId).isPresent();
         if (alreadyJoined) {
             throw new ChatException(ChatErrorCode.CHAT_ROOM_ALREADY_JOINED);
         }
@@ -80,10 +124,18 @@ public class ChatRoomService {
         findChatRoom.addUser(roomUser);
     }
 
-    public void leaveChatRoom(Long roomId, Long userId) {
-        ChatRoom findChatRoom = chatRoomRepository.findById(roomId)
+    public void leaveChatRoom(Long boardId, Long userId) {
+        boardRepository.findById(boardId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시글입니다."));
+
+        ChatRoom findChatRoom = chatRoomRepository.findByBoardId(boardId)
                 .orElseThrow(() -> new ChatException(ChatErrorCode.CHAT_ROOM_NOT_FOUND));
-        ChatRoomUser roomUser = chatRoomUserRepository.findByChatRoomIdAndUserId(roomId, userId)
+
+        if (findChatRoom.getType() == PRIVATE) {
+            throw new ChatException(ChatErrorCode.CHAT_ROOM_ACCESS_DENIED);
+        }
+
+        ChatRoomUser roomUser = chatRoomUserRepository.findByChatRoomIdAndUserId(findChatRoom.getId(), userId)
                 .orElseThrow(() -> new ChatException(ChatErrorCode.CHAT_ROOM_NOT_JOINED));
         chatRoomUserRepository.delete(roomUser);
         findChatRoom.removeUser(roomUser);
@@ -92,8 +144,10 @@ public class ChatRoomService {
     public void deleteChatRoom(Long roomId, Long userId) {
         ChatRoom chatRoom = chatRoomRepository.findChatRoomByIdAndUserId(roomId, userId)
                 .orElseThrow(() -> new ChatException(ChatErrorCode.CHAT_ROOM_DELETE_DENIED));
+
         ChatRoomUser roomUser = chatRoomUserRepository.findByChatRoomIdAndUserId(roomId, userId)
                         .orElseThrow(() -> new ChatException(ChatErrorCode.CHAT_ROOM_NOT_JOINED));
+
         if (roomUser.getRole().equals(ChatRole.MEMBER)) {
             throw new ChatException(ChatErrorCode.CHAT_ROOM_DELETE_DENIED);
         }
